+++ src/engine/Actions/Action/MoveBy.ts
@@ -34,7 +34,7 @@ export class MoveBy implements Action {
     }
   }
 
-  public update(_delta: number) {
+  public update(elapsedMs: number) {
     if (!this._started) {
       this._started = true;
       this._start = new Vector(this._tx.pos.x, this._tx.pos.y);