+++ src/engine/Collision/Detection/SparseHashGrid.ts
@@ -264,7 +264,7 @@ export class SparseHashGrid<TObject extends { bounds: BoundingBox }, TProxy exte
     return updated;
   }
 
-  debug(ex: ExcaliburGraphicsContext, delta: number): void {
+  debug(ex: ExcaliburGraphicsContext, elapsedMs: number): void {
     const transparent = Color.Transparent;
     const color = Color.White;
     for (const cell of this.sparseHashGrid.values()) {